Separate spending limit contract logic from safe contract logic
